Migrate convert package to generic FIX, WIP service package
