+++ service/websocket/fix42.go
@@ -1,6 +1,7 @@
 package websocket
 
 import (
+	"github.com/quickfixgo/field"
 	"strconv"
 
 	"github.com/bitfinexcom/bfxfixgw/convert"
@@ -54,7 +55,7 @@ func (w *Websocket) FIX42TradeHandler(t *bitfinex.Trade, sID quickfix.SessionID)
 		return nil
 	}
 	if reqID, ok := p.LookupMDReqID(t.Pair); ok {
-		fix := convert.FIX42MarketDataIncrementalRefreshFromTrade(reqID, t, w.Symbology, sID.TargetCompID)
+		fix := convert.FIXMarketDataIncrementalRefreshFromTrade(sID.BeginString, reqID, t, w.Symbology, sID.TargetCompID)
 		return quickfix.SendToTarget(fix, sID)
 	} else {
 		w.logger.Warn("could not find MDReqID for BFX trade", zap.String("Pair", t.Pair))
@@ -72,7 +73,7 @@ func (w *Websocket) FIX42TradeSnapshotHandler(s *bitfinex.TradeSnapshot, sID qui
 	if len(s.Snapshot) > 0 {
 		t := s.Snapshot[0]
 		if reqID, ok := p.LookupMDReqID(t.Pair); ok {
-			fix := convert.FIX42MarketDataFullRefreshFromTradeSnapshot(reqID, s, w.Symbology, sID.TargetCompID)
+			fix := convert.FIXMarketDataFullRefreshFromTradeSnapshot(sID.BeginString, reqID, s, w.Symbology, sID.TargetCompID)
 			return quickfix.SendToTarget(fix, sID)
 		} else {
 			w.logger.Warn("could not find MDReqID for BFX trade", zap.String("Pair", t.Pair))
@@ -119,7 +120,7 @@ func (w *Websocket) FIX42TradeExecutionUpdateHandler(t *bitfinex.TradeExecutionU
 	if err != nil {
 		return err
 	}
-	return quickfix.SendToTarget(convert.FIX42ExecutionReportFromTradeExecutionUpdate(t, p.BfxUserID(), cached.ClOrdID, cached.Qty, totalFillQty, cached.Px, cached.Stop, cached.Trail, avgFillPx, w.Symbology, sID.TargetCompID, cached.TifExpiration, cached.Flags), sID)
+	return quickfix.SendToTarget(convert.FIXExecutionReportFromTradeExecutionUpdate(sID.BeginString, t, p.BfxUserID(), cached.ClOrdID, cached.Qty, totalFillQty, cached.Px, cached.Stop, cached.Trail, avgFillPx, w.Symbology, sID.TargetCompID, cached.TifExpiration, cached.Flags), sID)
 }
 
 // FIX42BookSnapshot handles a book update snapshot
@@ -133,7 +134,7 @@ func (w *Websocket) FIX42BookSnapshot(s *bitfinex.BookUpdateSnapshot, sID quickf
 	if len(s.Snapshot) > 0 {
 		mdReqID, ok = p.LookupMDReqID(s.Snapshot[0].Symbol)
 		if ok {
-			return quickfix.SendToTarget(convert.FIX42MarketDataFullRefreshFromBookSnapshot(mdReqID, s, w.Symbology, sID.TargetCompID), sID)
+			return quickfix.SendToTarget(convert.FIXMarketDataFullRefreshFromBookSnapshot(sID.BeginString, mdReqID, s, w.Symbology, sID.TargetCompID), sID)
 		} else {
 			w.logger.Warn("could not find MDReqID for symbol", zap.String("MDReqID", mdReqID))
 		}
@@ -150,7 +151,7 @@ func (w *Websocket) FIX42BookUpdate(u *bitfinex.BookUpdate, sID quickfix.Session
 	}
 	mdReqID, ok := p.LookupMDReqID(u.Symbol)
 	if ok {
-		return quickfix.SendToTarget(convert.FIX42MarketDataIncrementalRefreshFromBookUpdate(mdReqID, u, w.Symbology, sID.TargetCompID), sID)
+		return quickfix.SendToTarget(convert.FIXMarketDataIncrementalRefreshFromBookUpdate(sID.BeginString, mdReqID, u, w.Symbology, sID.TargetCompID), sID)
 	} else {
 		w.logger.Warn("could not find MDReqID for symbol", zap.String("MDReqID", mdReqID))
 	}
@@ -178,7 +179,7 @@ func (w *Websocket) FIX42NotificationHandler(d *bitfinex.Notification, sID quick
 			if err == nil {
 				cxlClOrdID = cache.ClOrdID
 			}
-			return quickfix.SendToTarget(convert.FIX42OrderCancelReject(p.BfxUserID(), orderID, origClOrdID, cxlClOrdID, d.Text, false), sID)
+			return quickfix.SendToTarget(convert.FIXOrderCancelReject(sID.BeginString, p.BfxUserID(), orderID, origClOrdID, cxlClOrdID, d.Text, false), sID)
 		} else if d.Status == "SUCCESS" {
 			clOrdID := strconv.FormatInt(o.CID, 10)
 			orig, err := p.LookupByClOrdID(clOrdID)
@@ -188,9 +189,9 @@ func (w *Websocket) FIX42NotificationHandler(d *bitfinex.Notification, sID quick
 			}
 
 			exp, _ := convert.MTSToTime(orig.TifExpiration)
-			er := convert.FIX42ExecutionReport(orig.Symbol, orig.ClOrdID, orig.OrderID, orig.Account, enum.ExecType_PENDING_CANCEL, orig.Side, orig.Qty, 0.0, orig.FilledQty(), orig.Px, orig.Stop, orig.Trail, orig.AvgFillPx(), enum.OrdStatus_PENDING_CANCEL, orig.OrderType, orig.IsMargin, orig.TimeInForce, exp, d.Text, w.Symbology, sID.TargetCompID, orig.Flags)
+			er := convert.FIXExecutionReport(sID.BeginString, orig.Symbol, orig.ClOrdID, orig.OrderID, orig.Account, enum.ExecType_PENDING_CANCEL, orig.Side, orig.Qty, 0.0, orig.FilledQty(), orig.Px, orig.Stop, orig.Trail, orig.AvgFillPx(), enum.OrdStatus_PENDING_CANCEL, orig.OrderType, orig.IsMargin, orig.TimeInForce, exp, d.Text, w.Symbology, sID.TargetCompID, orig.Flags)
 			if orig.Px > 0 {
-				er.SetPrice(decimal.NewFromFloat(orig.Px), 4)
+				er.Set(field.NewPrice(decimal.NewFromFloat(orig.Px), 4))
 			}
 			return quickfix.SendToTarget(er, sID)
 		}
@@ -237,7 +238,7 @@ func (w *Websocket) FIX42NotificationHandler(d *bitfinex.Notification, sID quick
 			}
 			stop = o.Price
 		}
-		return quickfix.SendToTarget(convert.FIX42ExecutionReportFromOrder(&order, p.BfxUserID(), execType, 0, ordStatus, text, w.Symbology, sID.TargetCompID, flags, stop, peg), sID)
+		return quickfix.SendToTarget(convert.FIXExecutionReportFromOrder(sID.BeginString, &order, p.BfxUserID(), execType, 0, ordStatus, text, w.Symbology, sID.TargetCompID, flags, stop, peg), sID)
 	default:
 		w.logger.Warn("unhandled notify info object", zap.Any("msg", d.NotifyInfo))
 	}
@@ -275,8 +276,8 @@ func (w *Websocket) FIX42OrderSnapshotHandler(os *bitfinex.OrderSnapshot, sID qu
 				}
 			}
 			cache.OrderID = strconv.FormatInt(order.ID, 10)
-			er := convert.FIX42ExecutionReportFromOrder(order, peer.BfxUserID(), enum.ExecType_NEW, cache.FilledQty(), enum.OrdStatus_NEW, string(order.Status), w.Symbology, sID.TargetCompID, int(order.Flags), order.PriceAuxLimit, order.PriceTrailing)
-			er.SetAvgPx(decimal.NewFromFloat(cache.AvgFillPx()), 4)
+			er := convert.FIXExecutionReportFromOrder(sID.BeginString, order, peer.BfxUserID(), enum.ExecType_NEW, cache.FilledQty(), enum.OrdStatus_NEW, string(order.Status), w.Symbology, sID.TargetCompID, int(order.Flags), order.PriceAuxLimit, order.PriceTrailing)
+			er.Set(field.NewAvgPx(decimal.NewFromFloat(cache.AvgFillPx()), 4))
 			return quickfix.SendToTarget(er, sID)
 		}
 	}
@@ -312,7 +313,7 @@ func (w *Websocket) FIX42OrderUpdateHandler(o *bitfinex.OrderUpdate, sID quickfi
 			peg = cache.Trail
 		}
 	}
-	return quickfix.SendToTarget(convert.FIX42ExecutionReportFromOrder(&ord, p.BfxUserID(), execType, 0.0, ordStatus, "", w.Symbology, sID.TargetCompID, int(o.Flags), stop, peg), sID)
+	return quickfix.SendToTarget(convert.FIXExecutionReportFromOrder(sID.BeginString, &ord, p.BfxUserID(), execType, 0.0, ordStatus, "", w.Symbology, sID.TargetCompID, int(o.Flags), stop, peg), sID)
 }
 
 //FIX42OrderCancelHandler handles order cancels
@@ -337,5 +338,5 @@ func (w *Websocket) FIX42OrderCancelHandler(o *bitfinex.OrderCancel, sID quickfi
 	if ordStatus == enum.OrdStatus_FILLED || ordStatus == enum.OrdStatus_PARTIALLY_FILLED {
 		return nil // do not publish duplicate execution report--tu/te will have more information (fees, etc.) for this event
 	}
-	return quickfix.SendToTarget(convert.FIX42ExecutionReportFromOrder(&ord, p.BfxUserID(), execType, cached.FilledQty(), ordStatus, string(ord.Status), w.Symbology, sID.TargetCompID, cached.Flags, 0.0, cached.Trail), sID)
+	return quickfix.SendToTarget(convert.FIXExecutionReportFromOrder(sID.BeginString, &ord, p.BfxUserID(), execType, cached.FilledQty(), ordStatus, string(ord.Status), w.Symbology, sID.TargetCompID, cached.Flags, 0.0, cached.Trail), sID)
 }
